+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - OpenGL queries checks Point Size Range
 - Added functions to convert material float arrays into a glm object
 - Use glm::value_ptr() instead of &[0] to access beginning of data array
+- Query all available OpenGL extensions
+  - Display the list in OpenGLUtils when printing info
+  - Store a set of available extensions in OpenGLEngine to do a quick boolean query to check if a specific extension exists
 
 ## v 5.1.0 - 10 May 2024
 - Added CSCI441::drawSolidDome() and CSCI441::drawWireDome() which is a half-sphere with an open bottom